@@ -5,25 +5,29 @@
     "npm": "please-use-yarn"
   },
   "scripts": {
-    "start": "node dist/index.js",
-    "dev": "nodemon dist/index.js",
-    "start-ts": "ts-node src/index.ts",
-    "dev-ts": "nodemon --exec ts-node src/index.ts",
-    "watch": "tsc -w",
+    "start": "npm run build:dev && npm run server:dev",
+    "build:prod": "webpack --env.NODE_ENV=prod",
+    "build:dev": "webpack --env.NODE_ENV=dev",
+    "server:dev": "nodemon \"./dist/server.dev.js\" --watch \"./dist/server.dev.js\"",
+    "watch:dev": "webpack --env.NODE_ENV=dev --watch",
+    "server:prod": "nodemon \"./dist/server.prod.js\" --watch \"./dist/server.prod.js\"",
     "prepare": "husky install"
   },
   "dependencies": {
     "argon2": "^0.28.7",
     "cors": "^2.8.5",
     "dayjs": "^1.11.4",
-    "dotenv": "^16.0.1",
-    "express": "^4.18.1",
+    "dotenv": "^16.3.1",
+    "dotenv-webpack": "^8.0.1",
+    "express": "^4.18.2",
     "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.21",
     "mongodb": "^4.8.0",
     "nanoid": "^4.0.0",
     "pino": "^8.2.0",
     "pino-pretty": "^8.1.0",
+    "ts-loader": "^6.0.1",
+    "webpack": "^4.32.2",
     "zod": "^3.17.9"
   },
   "devDependencies": {
@@ -39,12 +43,16 @@
     "@types/nanoid": "^3.0.0",
     "@types/node": "^18.0.6",
     "@types/pino": "^7.0.5",
+    "@types/webpack": "^4.4.32",
+    "clean-webpack-plugin": "^3.0.0",
+    "concurrently": "^8.2.0",
     "husky": "^8.0.1",
     "lint-staged": "^13.0.3",
     "nodemon": "^2.0.19",
     "prettier": "^2.7.1",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.7.4"
+    "typescript": "^5.1.6",
+    "webpack-cli": "^3.3.2",
+    "webpack-node-externals": "^1.7.2"
   },
   "name": "nodejs_typescript",
   "version": "1.0.0",
